# backend/docs/swagger.yaml
basePath: /
definitions:
...
  entity.Payment:
    properties:
      amount:
        type: number
      clientID:
        type: string
      createdAt:
        type: string
      currency:
        type: string
      id:
        type: string
      masterID:
        type: string
      status:
        $ref: '#/definitions/entity.PaymentStatus'
      tonTransactionID:
        type: string
      type:
        $ref: '#/definitions/entity.PaymentType'
    type: object
  entity.PaymentStatus:
    enum:
    - pending
    - completed
    - failed
    type: string
    x-enum-varnames:
    - PaymentStatusPending
    - PaymentStatusCompleted
    - PaymentStatusFailed
  entity.PaymentType:
    enum:
    - payment
    - tip
    type: string
    x-enum-varnames:
    - PaymentTypePayment
    - PaymentTypeTip
...
  entity.Service:
    properties:
      categoryID:
        type: string
      createdAt:
        type: string
      description:
        type: string
      duration:
        type: string
      id:
        type: string
      photoURL:
        type: string
      price:
        type: number
      title:
        type: string
      userID:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a beauty services platform API server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: BeautyTON API
  version: "1.0"
paths:
...
  /payments:
    post:
      consumes:
      - application/json
      description: Create a new payment with the input payload
      parameters:
      - description: Create payment
        in: body
        name: payment
        required: true
        schema:
          $ref: '#/definitions/entity.Payment'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.Payment'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new payment
      tags:
      - payments
  /payments/{id}:
    get:
      consumes:
      - application/json
      description: Get payment details by payment ID
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Payment'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a payment by ID
      tags:
      - payments
    put:
      consumes:
      - application/json
      description: Update payment details by payment ID
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      - description: Update payment
        in: body
        name: payment
        required: true
        schema:
          $ref: '#/definitions/entity.Payment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Payment'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a payment
      tags:
      - payments
  /payments/{id}/status:
    put:
      consumes:
      - application/json
      description: Update payment status by payment ID
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      - description: Payment status
        in: body
        name: status
        required: true
        schema:
          properties:
            status:
              $ref: '#/definitions/entity.PaymentStatus'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Payment'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update payment status
      tags:
      - payments
...
  /services:
    post:
      consumes:
      - application/json
      description: Create a new service with the input payload
      parameters:
      - description: Create service
        in: body
        name: service
        required: true
        schema:
          $ref: '#/definitions/entity.Service'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.Service'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new service
      tags:
      - services
  /services/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a service by service ID
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a service
      tags:
      - services
    get:
      consumes:
      - application/json
      description: Get service details by service ID
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Service'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a service by ID
      tags:
      - services
    put:
      consumes:
      - application/json
      description: Update service details by service ID
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      - description: Update service
        in: body
        name: service
        required: true
        schema:
          $ref: '#/definitions/entity.Service'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Service'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a service
      tags:
      - services
  /services/{id}/photo:
    post:
      consumes:
      - multipart/form-data
      description: Upload a photo for a service
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      - description: Photo file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Upload service photo
      tags:
      - services
...
swagger: "2.0"